requirements.txt
asgiref==3.6.0
cmake==3.25.0
crispy-bootstrap5==0.7
Django==4.1
django-ckeditor==5.9.0
django-cors-headers==4.3.1
django-crispy-forms==2.0
django-js-asset==2.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
imutils==0.5.4
mysql==0.0.3
mysqlclient==2.1.1
numpy==1.24.1
opencv-contrib-python==4.10.0.82
opencv-python==4.10.0.82
Pillow==9.3.0
PyJWT==2.7.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
